move kubeconfig extension into own package
To overcome some import cycles, this commit moves the kubeconfig extension type and also functions related to the kubeconfig into its own package.
