+++ get/clusters.go
@@ -9,7 +9,7 @@ import (
 
 	infrastructure "github.com/ninech/apis/infrastructure/v1alpha1"
 	"github.com/ninech/nctl/api"
-	"github.com/ninech/nctl/auth"
+	"github.com/ninech/nctl/api/config"
 	"github.com/ninech/nctl/internal/format"
 )
 
@@ -38,7 +38,7 @@ func (l *clustersCmd) Run(ctx context.Context, client *api.Client, get *Cmd) err
 		return format.PrettyPrintObjects(clusterList.GetItems(), format.PrintOpts{})
 	case contexts:
 		for _, cluster := range clusterList.Items {
-			fmt.Printf("%s\n", auth.ContextName(&cluster))
+			fmt.Printf("%s\n", config.ContextName(&cluster))
 		}
 	}
