+++ get/project_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	management "github.com/ninech/apis/management/v1alpha1"
-	"github.com/ninech/nctl/auth"
+	"github.com/ninech/nctl/api/config"
 	"github.com/ninech/nctl/internal/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -148,5 +148,5 @@ func TestProjectsConfigErrors(t *testing.T) {
 	kubeconfig, err := test.CreateTestKubeconfig(apiClient, "")
 	require.NoError(t, err)
 	defer os.Remove(kubeconfig)
-	require.ErrorIs(t, cmd.Run(ctx, apiClient, get), auth.ErrConfigNotFound)
+	require.ErrorIs(t, cmd.Run(ctx, apiClient, get), config.ErrExtensionNotFound)
 }
